Moved serverstats embed and created function
